Move functions from utils.rs to utils/receiving.rs and utils/hash.rs
